requirements.txt
asgiref==3.8.1
asttokens==3.0.0
Bottleneck==1.4.2
certifi==2024.8.30
cftime==1.6.4.post1
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dask==2024.12.0
decorator==5.1.1
deprecation==2.1.0
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.1.0
fonttools==4.55.3
fsspec==2024.10.0
geopandas==1.0.1
gurobipy==12.0.0
highspy==1.8.1
importlib_metadata==8.5.0
ipython==8.30.0
jedi==0.19.2
kiwisolver==1.4.7
linopy==0.4.3
locket==1.0.0
matplotlib==3.9.4
matplotlib-inline==0.1.7
netCDF4==1.7.2
networkx==3.4.2
numexpr==2.10.2
numpy==1.26.4
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
partd==1.4.2
pillow==11.0.0
polars==1.17.1
prompt_toolkit==3.0.48
pure_eval==0.2.3
Pygments==2.18.0
PyJWT==2.10.1
pyogrio==0.10.0
pyparsing==3.2.0
pyproj==3.7.0
pypsa==0.32.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
scipy==1.14.1
shapely==2.0.6
six==1.17.0
sqlparse==0.5.2
stack-data==0.6.3
toolz==1.0.0
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
validators==0.34.0
wcwidth==0.2.13
xarray==2024.11.0
zipp==3.21.0